Create addition code
In this program, user is asked to enter two integers. Then, the sum of those two integers is stored in a variable and displayed on the screen. Primary tabs n this program, user is asked to enter two integers. These two integers are stored in variables firstNumber and secondNumber respectively.
Then, the variables firstNumberand secondNumber are added using + operator and stored in sumOfTwoNumbers variable.
Finally, sum Of Two Numbers is displayed on the screen.