Create to find factorial of the given number
In this program, we take a positive integer from the user and compute the factorial using for loop. We print an error message if the user enters a negative number.

We declare the type of factorial variable as long double since the factorial of a number may be very large.

When the user enters a positive integer (say 4), for loop is executed and computes the factorial. The value of i is initially 1.

The program runs until the statement i <= n becomes false. This prints Factorial of 4 = 24 on the screen. Here’s how the program executes when n = 4.