Update for beginners
A "Hello, World!" is a simple program that outputs Hello, World! on the screen. Since it's a very simple program, it's often used to introduce a new programming language to a newbie.

Let's see how C++ "Hello, World!" program works.

If you haven't already set up the environment to run C++ on your computer, visit Install C++ on Your Computer.